move CircularBuffer to nupic.utils
